some refactor shenanigans
